Subclass of RulerUpdater, common base of non-custom updaters
