+++ src/CMakeLists.txt
@@ -1030,6 +1030,8 @@ list( APPEND SOURCES
       >
       widgets/FileHistory.cpp
       widgets/FileHistory.h
+      widgets/GeneratedUpdater.cpp
+      widgets/GeneratedUpdater.h
       widgets/Grabber.cpp
       widgets/Grabber.h
       widgets/Grid.cpp
