+++ src/widgets/RulerUpdater.cpp
@@ -544,58 +544,6 @@ auto RulerUpdater::MakeTick(
    return { { strLeft, strTop, strW, strH }, lab };
 }
 
-bool RulerUpdater::Tick(wxDC& dc,
-   int pos, double d, const TickSizes& tickSizes, wxFont font,
-   // in/out:
-   TickOutputs outputs,
-   const RulerStruct& context) const
-{
-   const double mDbMirrorValue = context.mDbMirrorValue;
-   const int mLength = context.mLength;
-   const RulerFormat mFormat = context.mFormat;
-
-   const int mLeft = context.mLeft;
-   const int mTop = context.mTop;
-   const int mBottom = context.mBottom;
-   const int mRight = context.mRight;
-   const int mOrientation = context.mOrientation;
-
-   const RulerStruct::Fonts& mFonts = *context.mpFonts;
-   const TranslatableString mUnits = context.mUnits;
-   const int mSpacing = context.mSpacing;
-   const bool mFlip = context.mFlip;
-
-   // Bug 521.  dB view for waveforms needs a 2-sided scale.
-   if ((mDbMirrorValue > 1.0) && (-d > mDbMirrorValue))
-      d = -2 * mDbMirrorValue - d;
-
-   // FIXME: We don't draw a tick if off end of our label arrays
-   // But we shouldn't have an array of labels.
-   if (outputs.labels.size() >= mLength)
-      return false;
-
-   Label lab;
-   lab.value = d;
-   lab.pos = pos;
-   lab.text = tickSizes.LabelString(d, mFormat);
-   lab.units = mUnits;
-
-   const auto result = MakeTick(
-      lab,
-      dc, font,
-      outputs.bits,
-      mLeft, mTop, mSpacing, mFonts.lead,
-      mFlip,
-      mOrientation);
-
-   auto& rect = result.first;
-   outputs.box.Union(rect);
-   outputs.labels.emplace_back(result.second);
-   return !rect.IsEmpty();
-}
-
-RulerUpdater::~RulerUpdater() = default;
-
 // Formerly the ending part of Ruler::Updater::Update after a switch
 // on mCustom and mLog; now a base class function used in
 // overrides of RulerUpdater::Update as the common ending step
@@ -648,3 +596,5 @@ const
    for (auto& label : allOutputs.minorMinorLabels)
       update(label);
 }
+
+RulerUpdater::~RulerUpdater() = default;
